test(esl-listeners-test): move once subscription test
